requirements.txt
Flask
requests
gunicorn
psycopg2-binary
python-dotenv
